refactor(organization): extract EditOrganizationFeature component and simplify edit page
