+++ features/organization/index.ts
@@ -1,3 +1,4 @@
+export { EditOrganizationFeature } from './components/edit-organization-feature';
 export { OrganizationCard } from './components/organization-card';
 export { OrganizationForm } from './components/organization-form';
 export { OrganizationSelector } from './components/organization-selector';
